requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
dj-database-url>=2.1
psycopg[binary]>=3.1
django-htmx>=1.17
celery[redis]>=5.3
redis>=5.0
pydantic>=2.0
openpyxl>=3.1
django-simple-history>=3.4
django-filter>=24.0
whitenoise>=6.6
gunicorn>=22.0
Markdown>=3.5
bleach>=6.1
